use battery count circle
